show three buttons on new/empty notebook
As part of this change, moved to using the editor-cell template to render the buttons, and register for two separate events for above/below instead of multiplexing.

Note: there are three ways to get to an empty notebook - creating a new one, saving an empty one, and deleting your current notebook down to empty.